Improve functional consistency.
Previously, there was little distinction between models and blocks. In reality, a block
is a separate construct which represents a chunk of related lines within an SOC script.

To fully decouple models (i.e. `Thing`, `Level`) from script details, parsing logic has
been completely removed from model classes. Instead, parsing logic now exists in a
corresponding `Block` subclass for each corresponding model. Just as with `PropertyLines`,
the `unapply` of a model's block class (i.e. `ThingBlock`) is used to produce instances
of that model (i.e. `Thing`).

This has the added benefit of improving the readability of the parsing logic in `SocScript`
used to identify chunks of related lines as blocks (just as with `PropertyLines`, we can
use a `match` expression).

Similarly, the plan is to implement re-serialization of models to SOC script block lines
as the `apply` of corresponding model blocks. You "apply" a block function obj to a model
to get its SOC script representation.