Implement proper porting rules, with support for user-supplied prefix.
When in port mode, adjusted IDs will be prefixed with as much of the requested prefix as possible.

For example, the following transformations would occur by specifying '--freeslot-prefix Ghoul' on these resources:

Thing 361 => Thing MT_Ghoul361
State 2048 => State S_Ghoul2048
Sound 156 => sfx_gho156
Sprite 230 => SPR_GH6E

Note that in all cases except for Sprites, the original ID is preserved for readibility. For sprites, this ID is preserved but in base36 (230 base10 => 6E base36), because sprite IDs are limited to 4 characters.