Improve default behavior of sprite freeslot porting.
Previously, all sprite references would be upgraded, even if the state referencing them was not also being upgraded to a freeslot.

This behavior seemed more likely to be problematic, since a state already making use of freeslots is likely depending on some hard-coded ID on purpose.

Now, the upgrade is only performed if at least one state referencing the hard-coded sprite slot is also being upgraded.