Fix sprite dependency listing.
We now consider any sprite that's not declared as a freeslot to be an external reference. This implies that before porting
a SOC which uses a hard-coded slot number for a sprite will be considered to have have an external sprite reference on that
slot number. However, if the SOC is ported, that sprite number will be upgraded to a freeslot, at which point it'll no
longer be considered an external reference.